+++ frontend/src/features/crawls/index.ts
@@ -2,3 +2,4 @@ import("./crawl-list");
 import("./crawl-state-filter");
 import("./crawler-channel-badge");
 import("./proxy-badge");
+import("./url-list");
